requirements.txt
flask
flask_injector
google_cloud_storage